updated collision mechanic

+++ SuperStory/src/ECS/ColliderComponent.h
@@ -32,8 +32,8 @@ public:
 	{
 		collider.x = static_cast<int>(transform->position.x);
 		collider.y = static_cast<int>(transform->position.y);
-		collider.w = static_cast<int>(transform->width * transform->scale);
-		collider.h = static_cast<int>(transform->height * transform->scale);
+		collider.w = transform->width * transform->scale;
+		collider.h = transform->height * transform->scale;
 	}
 
 };
